+++ aiohomematic/model/generic/data_point.py
@@ -16,7 +16,7 @@ from aiohomematic import i18n
 from aiohomematic.const import DP_KEY_VALUE, DataPointUsage, EventType, Parameter, ParameterData, ParamsetKey
 from aiohomematic.decorators import inspector
 from aiohomematic.exceptions import ValidationException
-from aiohomematic.interfaces import ChannelProtocol, GenericDataPointProtocol
+from aiohomematic.interfaces.model import ChannelProtocol, GenericDataPointProtocol
 from aiohomematic.model import data_point as hme
 from aiohomematic.model.support import DataPointNameData, get_data_point_name_data
 from aiohomematic.property_decorators import hm_property