+++ aiohomematic/model/week_profile.py
@@ -380,10 +380,10 @@ from aiohomematic.const import (
 )
 from aiohomematic.decorators import inspector
 from aiohomematic.exceptions import ClientException, ValidationException
-from aiohomematic.interfaces import WeekProfileProtocol
+from aiohomematic.interfaces.model import WeekProfileProtocol
 
 if TYPE_CHECKING:
-    from aiohomematic.interfaces import CustomDataPointProtocol
+    from aiohomematic.interfaces.model import CustomDataPointProtocol
     from aiohomematic.model.custom import BaseCustomDpClimate
 
 _LOGGER = logging.getLogger(__name__)