+++ aiohomematic_test_support/factory.py
@@ -44,7 +44,7 @@ from aiohomematic.central import CentralConfig, CentralUnit
 from aiohomematic.central.event_bus import BackendSystemEventData, HomematicEvent
 from aiohomematic.client import ClientConfig, InterfaceConfig
 from aiohomematic.const import LOCAL_HOST, BackendSystemEvent, Interface, OptionalSettings
-from aiohomematic.interfaces import ClientProtocol
+from aiohomematic.interfaces.client import ClientProtocol
 from aiohomematic_test_support import const
 from aiohomematic_test_support.mock import SessionPlayer, get_client_session, get_mock, get_xml_rpc_proxy