+++ tests/conftest.py
@@ -12,7 +12,7 @@ import pytest
 
 from aiohomematic.central import CentralUnit
 from aiohomematic.central.event_bus import BackendSystemEventData, HomematicEvent
-from aiohomematic.interfaces import ClientProtocol
+from aiohomematic.interfaces.client import ClientProtocol
 from aiohomematic_test_support import const
 from aiohomematic_test_support.factory import (
     FactoryWithClient,